fix: fix BufferReader class
Fix BufferReader class to correctly handle the case where the buffer length is less than or equal to the offset.

+++ app/src/main/java/protobuf/magic/Part.java
@@ -1,14 +1,12 @@
 package protobuf.magic;
 
-import java.math.BigInteger;
-
 public class Part {
   int[] byteRange;
-  BigInteger index;
-  BigInteger type;
+  int index;
+  int type;
   Object value;
 
-  Part(int[] byteRange, BigInteger index, BigInteger type, Object value) {
+  Part(int[] byteRange, int index, int type, Object value) {
     this.byteRange = byteRange;
     this.index = index;
     this.type = type;
